+++ NEWS.md
@@ -1,3 +1,16 @@
+# survivoR 2.0.1
+
+New helper functions:
+
+* `show_palette` for viewing the season palette generated from the logo
+* `get_castaway_image` returns the URL of the castaway thumbnail for visualisation
+
+Data:
+
+* Updated US43 data
+* Updated vote history data
+
+
 # survivoR 2.0
 
 Big update!